Move tbot CA rotation test to integration.
